+++ src/app/app-config.ts
@@ -1,0 +1,85 @@
+import { Configuration } from 'msal';
+import { MsalAngularConfiguration } from '@azure/msal-angular';
+
+// this checks if the app is running on IE
+export const isIE = window.navigator.userAgent.indexOf('MSIE ') > -1 || window.navigator.userAgent.indexOf('Trident/') > -1;
+
+/** =================== REGIONS ====================
+ * 
+ * 1) Web API configuration parameters
+ * 2) Authentication configuration parameters
+ * 3) MSAL-Angular specific configuration parameters
+ * 
+ * ================================================= 
+*/ 
+
+
+
+// #region 1) Web API Configuration
+/** 
+ * Enter here the coordinates of your Web API and scopes for access token request
+ * The current application coordinates were pre-registered in a B2C tenant.
+ */
+export const apiConfig = {
+    b2cScopes: ['https://fabrikamb2c.onmicrosoft.com/helloapi/demo.read'],
+    webApi: 'https://fabrikamb2chello.azurewebsites.net/hello'
+};
+// #endregion
+
+
+
+// #region 2) Authentication Configuration
+/** 
+ * Config object to be passed to Msal on creation. For a full list of msal.js configuration parameters,
+ * visit https://azuread.github.io/microsoft-authentication-library-for-js/docs/msal/modules/_configuration_.html
+ */
+export const msalConfig: Configuration = {
+    auth: {
+        clientId: "e760cab2-b9a1-4c0d-86fb-ff7084abd902",
+        authority: "https://fabrikamb2c.b2clogin.com/fabrikamb2c.onmicrosoft.com/b2c_1_susi",
+        validateAuthority: false,
+        redirectUri: "http://localhost:6420/",
+      },
+    cache: {
+        cacheLocation: "localStorage",
+        storeAuthStateInCookie: isIE, // Set this to "true" to save cache in cookies to address trusted zones limitations in IE
+    },
+}
+
+/** 
+ * Scopes you enter here will be consented once you authenticate. For a full list of available authentication parameters, 
+ * visit https://azuread.github.io/microsoft-authentication-library-for-js/docs/msal/modules/_authenticationparameters_.html
+ */
+export const loginRequest = {
+    scopes: ['openid', 'profile'],
+};
+
+// Scopes you enter will be used for the access token request for your web API
+export const tokenRequest = {
+    scopes: apiConfig.b2cScopes // i.e. [https://fabrikamb2c.onmicrosoft.com/helloapi/demo.read]
+};
+// #endregion
+
+
+
+// #region 3) MSAL-Angular Configuration
+// here you can define the coordinates and required permissions for your protected resources
+export const protectedResourceMap: [string, string[]][] = [
+    [apiConfig.webApi, apiConfig.b2cScopes] // i.e. [https://fabrikamb2chello.azurewebsites.net/hello, ['https://fabrikamb2c.onmicrosoft.com/helloapi/demo.read']]
+];
+
+/** 
+ * MSAL-Angular specific authentication parameters. For a full list of available options,
+ * visit https://github.com/AzureAD/microsoft-authentication-library-for-js/tree/dev/lib/msal-angular#config-options-for-msal-initialization
+*/
+export const msalAngularConfig: MsalAngularConfiguration = {
+    popUp: !isIE,
+    consentScopes: [
+        ...loginRequest.scopes,
+        ...tokenRequest.scopes,
+    ],
+    unprotectedResources: [], // API calls to these coordinates will NOT activate MSALGuard
+    protectedResourceMap,     // API calls to these coordinates will activate MSALGuard
+    extraQueryParameters: {}  
+}
+// #endregion