+++ src/app/app.module.ts
@@ -1,18 +1,42 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { NgModule } from '@angular/core';
-import { MatToolbarModule, MatButtonModule, MatListModule, MatCardModule } from '@angular/material';
+import { HTTP_INTERCEPTORS, HttpClientModule } from '@angular/common/http';
+
+import { MatButtonModule } from '@angular/material/button';
+import { MatCardModule } from '@angular/material/card';
+import { MatListModule } from '@angular/material/list';
+import { MatToolbarModule } from '@angular/material/toolbar';
+
+import { Configuration } from 'msal';
+import {
+  MsalModule,
+  MsalInterceptor,
+  MSAL_CONFIG,
+  MSAL_CONFIG_ANGULAR,
+  MsalService,
+  MsalAngularConfiguration
+} from '@azure/msal-angular';
+
+import { msalConfig, msalAngularConfig } from './app-config';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { MsalModule, MsalInterceptor } from '@azure/msal-angular';
-import { HTTP_INTERCEPTORS, HttpClientModule } from '@angular/common/http';
-import { msalConfig, apiConfig, isIE } from './config';
 import { HomeComponent } from './home/home.component';
+import { ProfileComponent } from './profile/profile.component';
+
+function MSALConfigFactory(): Configuration {
+  return msalConfig;
+}
+
+function MSALAngularConfigFactory(): MsalAngularConfiguration {
+  return msalAngularConfig;
+}
 
 @NgModule({
   declarations: [
     AppComponent,
     HomeComponent,
+    ProfileComponent,
   ],
   imports: [
     BrowserModule,
@@ -22,18 +46,9 @@ import { HomeComponent } from './home/home.component';
     MatToolbarModule,
     MatButtonModule,
     MatListModule,
-    MatCardModule,
     AppRoutingModule,
-    MsalModule.forRoot(msalConfig,  
-    {
-      popUp: !isIE,
-      consentScopes: apiConfig.b2cScopes,
-      unprotectedResources: [],
-      protectedResourceMap: [
-          [apiConfig.webApi, apiConfig.b2cScopes],
-      ],
-      extraQueryParameters: {},
-    })
+    MatCardModule,
+    MsalModule
   ],
   providers: [
     {
@@ -41,6 +56,15 @@ import { HomeComponent } from './home/home.component';
       useClass: MsalInterceptor,
       multi: true
     },
+    {
+      provide: MSAL_CONFIG,
+      useFactory: MSALConfigFactory
+    },
+    {
+      provide: MSAL_CONFIG_ANGULAR,
+      useFactory: MSALAngularConfigFactory
+    },
+    MsalService
   ],
   bootstrap: [AppComponent]
 })