+++ src/app/home/home.component.ts
@@ -1,8 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { BroadcastService, MsalService} from '@azure/msal-angular';
 import { HttpClient } from '@angular/common/http';
-import { tokenRequest, isIE } from '../config';
-import { apiConfig } from '../config';
 
 @Component({
   selector: 'app-home',
@@ -11,8 +9,6 @@ import { apiConfig } from '../config';
 })
 export class HomeComponent implements OnInit {
   loggedIn = false;
-  tokenAcquired = false;
-  apiResponse: any;
 
   constructor(private broadcastService: BroadcastService, private authService: MsalService, private http: HttpClient) { }
 
@@ -25,12 +21,10 @@ export class HomeComponent implements OnInit {
 
     this.broadcastService.subscribe('msal:acquireTokenSuccess', (payload) => {
       console.log('access token acquired: ' + new Date().toString());
-      this.tokenAcquired = true;
     });
  
     this.broadcastService.subscribe('msal:acquireTokenFailure', (payload) => {
       console.log('access token acquisition fails');
-      this.tokenAcquired = false;
     });
 
     // redirect callback for redirect flow (IE)
@@ -41,7 +35,6 @@ export class HomeComponent implements OnInit {
       }
 
       console.log('Redirect Success: ', response.accessToken);
-      this.tokenAcquired = true;
     });
   }
 
@@ -49,27 +42,4 @@ export class HomeComponent implements OnInit {
   checkAccount() {
     this.loggedIn = !!this.authService.getAccount();
   }
-
-  getProfile() {
-    if (isIE) {
-      this.authService.loginRedirect(tokenRequest);
-    } else {
-      this.authService.loginPopup(tokenRequest)
-        .then(res => {
-          this.callAPI(apiConfig.webApi);
-        })
-        .catch(err => console.log(err));
-    }
-  }
-
-  callAPI(url: string) {
-    console.log('API call made at: ' + new Date().toString());
-
-    this.http.get(url).toPromise()
-      .then(res => {
-        this.apiResponse = res;
-        console.log('API responded at: ' + new Date().toString());
-      }).catch(err => console.log(err));
-  }
-
 }