+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # CHANGELOG
 
+
+## 3/31/2020
+
+* Updated Angular 8 to Angular 9.
+* Added a new route protected by MSALGuard.
+* Added Angular Factory for configuration objects.
+* Reorganized configuration file and added clarifications.
+* Minor changes on README.md.
+
+
 ## 2/20/2020
 
 * Initial version of sample application and related files.